+++ src/omop_cdm/dynamic/cdm54/derived_elements.py
@@ -23,8 +23,8 @@ class BaseConditionEraCdm54:
     condition_era_id: Mapped[int] = mapped_column(Integer, primary_key=True, sort_order=100)
     person_id: Mapped[int] = mapped_column(ForeignKey(FK_PERSON_ID, ondelete="CASCADE"), index=True, sort_order=200)
     condition_concept_id: Mapped[int] = mapped_column(ForeignKey(FK_CONCEPT_ID), index=True, sort_order=300)
-    condition_era_start_date: Mapped[datetime.datetime] = mapped_column(DateTime, sort_order=400)
-    condition_era_end_date: Mapped[datetime.datetime] = mapped_column(DateTime, sort_order=500)
+    condition_era_start_date: Mapped[datetime.date] = mapped_column(Date, sort_order=400)
+    condition_era_end_date: Mapped[datetime.date] = mapped_column(Date, sort_order=500)
     condition_occurrence_count: Mapped[Optional[int]] = mapped_column(Integer, sort_order=600)
 
     @declared_attr
@@ -46,8 +46,8 @@ class BaseDoseEraCdm54:
     drug_concept_id: Mapped[int] = mapped_column(ForeignKey(FK_CONCEPT_ID), index=True, sort_order=300)
     unit_concept_id: Mapped[int] = mapped_column(ForeignKey(FK_CONCEPT_ID), sort_order=400)
     dose_value: Mapped[decimal.Decimal] = mapped_column(Numeric, sort_order=500)
-    dose_era_start_date: Mapped[datetime.datetime] = mapped_column(DateTime, sort_order=600)
-    dose_era_end_date: Mapped[datetime.datetime] = mapped_column(DateTime, sort_order=700)
+    dose_era_start_date: Mapped[datetime.date] = mapped_column(Date, sort_order=600)
+    dose_era_end_date: Mapped[datetime.date] = mapped_column(Date, sort_order=700)
 
     @declared_attr
     def drug_concept(cls) -> Mapped["Concept"]:
@@ -70,8 +70,8 @@ class BaseDrugEraCdm54:
     drug_era_id: Mapped[int] = mapped_column(Integer, primary_key=True, sort_order=100)
     person_id: Mapped[int] = mapped_column(ForeignKey(FK_PERSON_ID, ondelete="CASCADE"), index=True, sort_order=200)
     drug_concept_id: Mapped[int] = mapped_column(ForeignKey(FK_CONCEPT_ID), index=True, sort_order=300)
-    drug_era_start_date: Mapped[datetime.datetime] = mapped_column(DateTime, sort_order=400)
-    drug_era_end_date: Mapped[datetime.datetime] = mapped_column(DateTime, sort_order=500)
+    drug_era_start_date: Mapped[datetime.date] = mapped_column(Date, sort_order=400)
+    drug_era_end_date: Mapped[datetime.date] = mapped_column(Date, sort_order=500)
     drug_exposure_count: Mapped[Optional[int]] = mapped_column(Integer, sort_order=600)
     gap_days: Mapped[Optional[int]] = mapped_column(Integer, sort_order=700)